#20 refactoring TopBar
