+++ app/src/main/java/com/example/myfulgora/ui/screens/tabs/map/MapScreen.kt
@@ -169,7 +169,6 @@ fun MapScreen(
                     .align(Alignment.TopCenter)
             ) {
                 FulgoraTopBar(
-                    userName = currentUser?.profile?.name ?: "Rider",
                     iconSize = iconSize,
                     onMenuClick = onMenuClick,
                     onUserClick = onUserClick,
